The cake service is the biggest change in `docker-compose.yml`. We added an `entrypoint` which changes the way the container works when we run it. Because we didn't define an ENTRYPOINT in our `Dockerfile`, we have some flexibility in how we run the application container. By default, we can run Apache, and then we can customize other containers to run the cake CLI.
According to the CakePHP 2 console documentation (https://book.cakephp.org/2.0/en/console-and-shells.html), you can pass the `-app` argument to customize the location of the application. We add it to the `entrypoint` just to be sure the Cake CLI knows how to locate the application.

We also added an `image:` key to both the `app` and `cake` services, so they run from the same image instead of building the same image twice. Using the container_name key on both services was added make the names of containers a little more comfortable on the eyes.

The other way we could run this without defining a docker-compose service is using the `--entrypoint` flag. You can start a container with bash and execute console commands within the container, which feels similar to SSH.
For example, `docker-compose run --entrypoint=/bin/bash app`