Retrofit our codebase to work with environment variables for configuration (e.g. DB password).
`vlucas/phpdotenv` supports our application's version of PHP! CakePHP 2.6 doesn't use Composer for its autoloading, but we can use it for adding the PHP dotenv package.

In CakePHP 2.6, the framework does not use Composer's autoloader, so we add the autoloader to the top of the entrypoint of the application.

Create our `.env` file, add an example `.env.example` file, and ignore the `.env` file in Git.

A system-defined variable takes precedence over a value in the `.env` file, and like modern applications, we can now set some defaults and override things with environment variables for production.

The PHP dotenv package doesn't perform type conversions on values, so `MY_ENV=true` is a string value `"true."`.
We borrow a little from the Laravel Support `helpers.php` file to make our integration support type conversions. We will provide some helper functions to read environment variables and autoload the file through composer's autoloader.
Our helper file defines two functions: `value()` and `cakeenv()`. The `value()` function allows the `cakeenv()` function to get the default value from a string or a Closure, and lucky for us, closures are supported as of >= PHP 5.3.0.

In order to get the helper loaded, we need to add an autoload key to our project's `composer.json` file.

`composer dump-autoload`
`docker-compose down`
`docker-compose up --build`