Use our new helpers to go back to the CakePHP warnings found when visiting http://localhost:8888 and update them with environment variables.
Find the configuration for `Security.salt` and `Security.cipherSeed` in the `app/Config/core.php` file and update them.
The default values we've set are the defaults that ship with CakePHP, which means that we'll have the same warnings when a new environment hasn't defined these values in an environment configuration.