Database Configuration
The database connections are properties of the `DATABASE_CONFIG` class. We need to update it slightly because we can't set the array properties with a dynamic `cakeenv()` function.

The CakePHP project ignores `/app/Config/database.php` for a good reason, but since we are using environment variables for connections, we can safely remove it from the `.gitignore` file and adapt the class.
We moved the public class properties to the constructor and set them dynamically. We didn't use the environment to change the test values, but it is possible to adjust those as well.

We've set the database environment variables in the container configuration to match the MySQL environment variables for the database and root password. These environment variables get defined on the system; therefore, PHP dotenv will not mutate these values (see https://github.com/vlucas/phpdotenv#immutability). What this means, is that you can override values found in the .env file by setting system environment variables.