Create a separate configuration file just for the CLI.
Suhosin: allow executing phar (Composer needs this to be installed).

The description of the simulation setting (https://suhosin.org/stories/configuration.html#suhosin-simulation) is defined as follows:
If you fear that Suhosin breaks your application, you can activate Suhosin's simulation mode with this flag. When Suhosin runs in simulation mode, violations are logged as usual, but nothing is blocked or removed from the request.
Using simulation mode is an excellent way to run your application and collect violations so that you can address them.

@TODO:
Another consideration we DID NOT cover is writing to the file system. You need to adjust the application to either start uploading files to Amazon S3 for example or persist uploads to a host machine with a volume.